requirements.txt
appdirs==1.4.4
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-email-verification==0.0.2
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.21
whitenoise==5.1.0
